requirements.txt
pydantic
gradio
google-genai